requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
qrcode==7.3
Pillow==8.3.1
python-dotenv==0.19.0
Flask-WTF==0.15.1
Flask-Migrate==3.1.0
requests==2.31.0
beautifulsoup4==4.12.2
isbnlib==3.10.14
python-amazon-paapi==4.1.0
